Create butane_cli lib.rs
